+++ packages/plugin-app-core/src/generator/templates/app/rax/TabBar/index.css
@@ -1,0 +1,31 @@
+.rax-app-tab-bar {
+  position: fixed;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  height: 98rpx;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  border: 1rpx solid #eee;
+  background-color: #fff;
+}
+
+.rax-app-tab-bar-item {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.rax-app-tab-bar-item-text {
+  font-size: 24rpx;
+}
+
+.rax-app-tab-bar-item-img {
+  margin-bottom: 8rpx;
+  width: 30rpx;
+  height: 30rpx;
+}
